requirements.txt
# Core Dependencies
python-telegram-bot==22.6
requests==2.32.5
python-dotenv==1.2.1
aiohttp==3.13.2
pytz==2025.2

# Database
supabase==2.28.0

# Blockchain
web3==7.6.0

# Testing
pytest==8.3.4
hypothesis==6.122.3

# Optional but recommended
certifi>=2023.7.22
urllib3>=2.0.0

# Note: PyDrive2 not needed anymore in V2.0
# We use Google Drive for Desktop (G: drive mount) instead
# Much simpler and more reliable!
